Various registration updates
Fixed registrationWizard, moved class and race const into types.ts so its not imported by client-side code, created a signup route and page + updatd the header links
