+++ src/lib/types.ts
@@ -1,12 +1,33 @@
-import type { CreatureRaceType, CreatureClassType } from "$lib/server/db/schema";
+export const CreatureClass = {
+	WARRIOR: 'warrior',
+	BRAWLER: 'brawler',
+	WIZARD: 'wizard',
+	CLERIC: 'cleric',
+	ASSASSIN: 'assassin',
+	ARCHER: 'archer',
+	ALCHEMIST: 'alchemist',
+	ENGINEER: 'engineer'
+  } as const;
+
+
+export const CreatureRace = {
+	HUMAN: 'human',
+	ORC: 'orc',
+	ELF: 'elf',
+	DWARF: 'dwarf',
+  } as const;
+
+export type CreatureClassType = typeof CreatureClass[keyof typeof CreatureClass];
+export type CreatureRaceType = typeof CreatureRace[keyof typeof CreatureRace];
 export interface RegistrationData {
+    email: string;
     username: string;
-    displayName: string;
-    bio: string;
-    preferences: {
-      notifications: boolean;
-      publicProfile: boolean;
+    password: string;
+    confirmPassword: string;
+    age: number;
+    creature: {
+      name: string;
+      class: CreatureClassType;
+      race: CreatureRaceType;
     };
-    class: CreatureClassType;
-    race: CreatureRaceType;
   }
